simplify pyproject.toml
And be more explicit about included package data:

- make sure that manpage is included in the sdist tar archive
- include tests in sdist archive but not in the wheel
- eliminate deprecation warnings in newer setuptools versions

Since the build.yaml workflow explicitly installs the built system_calls
wheel, there is no need to add the current working directory (CWD) to
pytest's pythonpath.

Furthermore, in case a future update breaks the system_calls wheel
creation, the pytest invocation will not detect it,
because it imports the package from the working directory and
not from the installed location.

FWIW, in case you want to explicitly run the tests against the CWD,
e.g. during development, you can invoke pytest like this from the CWD:

    python3 -m pytest -v

This works because it adds `.` to the pythonpath, by default,
in contrast to the `pytest ...` frontend.

closes #131